// studio-mss-portfolio/sanity.types.ts
export type Experience = {
  _id: string
  _type: 'experience'
  _createdAt: string
  _updatedAt: string
  _rev: string
  company?: string
  location?: string
  jobTitle?: string
  startDate?: string
  endDate?: string
  responsibilities?: Array<string>
}